int overflow

+++ src/main/java/com/actiontech/dble/DbleServer.java
@@ -325,7 +325,7 @@ public final class DbleServer {
         short bufferPoolPageNumber = system.getBufferPoolPageNumber();
         //minimum allocation unit
         short bufferPoolChunkSize = system.getBufferPoolChunkSize();
-        if (bufferPoolPageSize * bufferPoolPageNumber > Platform.getMaxDirectMemory()) {
+        if ((long) bufferPoolPageSize * (long) bufferPoolPageNumber > Platform.getMaxDirectMemory()) {
             throw new IOException("Direct BufferPool size[bufferPoolPageSize(" + bufferPoolPageSize + ")*bufferPoolPageNumber(" + bufferPoolPageNumber + ")] lager than MaxDirectMemory[" + Platform.getMaxDirectMemory() + "]");
         }
         bufferPool = new DirectByteBufferPool(bufferPoolPageSize, bufferPoolChunkSize, bufferPoolPageNumber);
